Plugin simplified
The plugin itself is now only 16.7 KB.

The Bible folder is now separate and should be placed in the
plugins/Tadukoo Bible folder. If it is not there, the plugin will not
work correctly. The books list will continue to have different colors
until all the books are completed in yml files. In game books will not
be available for a while.